hbl: Make hbl.py compatible with API
